Replace Sequential with Sequential2
Their performance are so close that it doesn't make sense to keep
both variants.
